better info for non-installed cores
